+++ lib/crossbeams/layout/fold_up.rb
@@ -81,8 +81,8 @@ module Crossbeams
 
         row_renders = nodes.reject(&:invisible?).map(&:render).join("\n")
         <<~HTML
-          <details class="pointer pv2"#{open_state}>
-            <summary class="b blue shadow-3 pa1 mr2">#{caption_text}</summary>
+          <details class="pv2"#{open_state}>
+            <summary class="pointer b blue shadow-3 pa1 mr2">#{caption_text}</summary>
             #{row_renders}
           </details>
         HTML